printing second half of string

+++ 0x05-pointers_arrays_strings/7-puts_half.c
@@ -8,7 +8,7 @@
 
 void puts_half(char *str)
 {
-	size_t length, start_index, i;
+	int length, start_index, i;
 
 	length = 0;
 	while (str[length] != '\0')
